feat: unify analysis job infrastructure by replacing individual model jobs with a single foldrun-analysis-job and shared utilities.
