+++ applications/foldrun/foldrun-agent/.env.example
@@ -97,13 +97,12 @@ GEMINI_MODEL=gemini-3-flash-preview
 # Format: {region}-docker.pkg.dev/{project}/{repo}/openfold3-components:latest
 # OPENFOLD3_COMPONENTS_IMAGE=us-central1-docker.pkg.dev/my-project-id/foldrun-repo/openfold3-components:latest
 
-# OF3 Analysis Job name (Cloud Run Job for parallel analysis)
-# OF3_ANALYSIS_JOB_NAME=of3-analysis-job
+# Analysis Job name (Cloud Run Job for parallel analysis)
+# ANALYSIS_JOB_NAME=foldrun-analysis-job
 
 # Optional: Boltz-2 integration
 # If not set, Boltz-2 tools are not loaded.
 # BOLTZ2_COMPONENTS_IMAGE=us-central1-docker.pkg.dev/my-project-id/foldrun-repo/boltz2-components:latest
-# BOLTZ2_ANALYSIS_JOB_NAME=boltz2-analysis-job
 # BOLTZ2_CACHE_PATH=boltz2/cache  # NFS path containing boltz2_conf.ckpt + mols/ (default: boltz2/cache)
 # ============================================================================
 # GPU Configuration (OPTIONAL)
