+++ applications/foldrun/foldrun-agent/foldrun_app/models/boltz2/tools/analyze_job.py
@@ -32,7 +32,7 @@ class BOLTZ2JobAnalysisTool(BOLTZ2Tool):
 
     def __init__(self, tool_config: Dict[str, Any], config: Any):
         super().__init__(tool_config, config)
-        self.job_name = os.getenv("BOLTZ2_ANALYSIS_JOB_NAME", "boltz2-analysis-job")
+        self.job_name = os.getenv("ANALYSIS_JOB_NAME", "foldrun-analysis-job")
 
     def _get_analysis_path(self, job: Any) -> str:
         """Get GCS analysis path for a job."""
@@ -260,6 +260,7 @@ class BOLTZ2JobAnalysisTool(BOLTZ2Tool):
                         run_v2.RunJobRequest.Overrides.ContainerOverride(
                             env=[
                                 run_v2.EnvVar(name="ANALYSIS_PATH", value=analysis_path),
+                                run_v2.EnvVar(name="MODEL_TYPE", value="boltz2"),
                                 run_v2.EnvVar(
                                     name="GEMINI_MODEL",
                                     value=os.getenv(
