--- applications/foldrun/src/boltz2-analysis-job/requirements.txt
@@ -1,7 +1,0 @@
-google-cloud-storage>=3.0.0
-google-cloud-aiplatform>=1.85.0
-google-genai>=1.0.0
-numpy>=2.2.3
-matplotlib>=3.10.0
-seaborn>=0.13.2
-pyyaml>=6.0
