+++ applications/foldrun/src/foldrun-analysis-job/cloudbuild.yaml
