+++ applications/foldrun/src/foldrun-analysis-job/of3_analyzer.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # Copyright 2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,17 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Cloud Run Job for analyzing OpenFold3 predictions in parallel.
-
-OF3 outputs JSON (not pickle like AF2), so this job is lighter:
-no jax, no dm-tree, no pickle. Reads *_confidences.json and
-*_confidences_aggregated.json to produce plots and Gemini analysis.
-"""
+"""OpenFold3 specific prediction analyzer and consolidator."""
 
 import json
 import logging
 import os
 import sys
+import time
 from datetime import datetime, timezone
 
 import matplotlib
@@ -36,133 +31,20 @@ matplotlib.use("Agg")  # Use non-interactive backend for Cloud Run
 import matplotlib.pyplot as plt
 import seaborn as sns
 
-# Configure logging
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+from shared_utils import (
+    calculate_plddt_stats,
+    download_json_from_gcs,
+    download_text_from_gcs,
+    get_quality_assessment,
+    upload_to_gcs,
 )
+
 logger = logging.getLogger(__name__)
-
-# pLDDT confidence bands (same as AF2)
-PLDDT_BANDS = [
-    (0, 50, "very_low_confidence"),
-    (50, 70, "low_confidence"),
-    (70, 90, "high_confidence"),
-    (90, 100, "very_high_confidence"),
-]
-
-
-def download_from_gcs(gcs_uri: str, local_path: str) -> None:
-    """Download file from GCS."""
-    if not gcs_uri.startswith("gs://"):
-        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
-
-    parts = gcs_uri[5:].split("/", 1)
-    bucket_name = parts[0]
-    blob_name = parts[1] if len(parts) > 1 else ""
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(blob_name)
-    blob.reload()
-    blob.download_to_filename(local_path)
-
-    size_mb = blob.size / 1024 / 1024 if blob.size else 0
-    logger.info(f"Downloaded {gcs_uri} ({size_mb:.2f} MB)")
-
-
-def upload_to_gcs(local_path: str, gcs_uri: str) -> None:
-    """Upload file to GCS."""
-    if not gcs_uri.startswith("gs://"):
-        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
-
-    parts = gcs_uri[5:].split("/", 1)
-    bucket_name = parts[0]
-    blob_name = parts[1] if len(parts) > 1 else ""
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(blob_name)
-    blob.upload_from_filename(local_path)
-
-    logger.info(f"Uploaded {local_path} to {gcs_uri}")
-
-
-def download_json_from_gcs(gcs_uri: str) -> dict:
-    """Download and parse JSON file from GCS."""
-    if not gcs_uri.startswith("gs://"):
-        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
-
-    parts = gcs_uri[5:].split("/", 1)
-    bucket_name = parts[0]
-    blob_name = parts[1] if len(parts) > 1 else ""
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(blob_name)
-
-    content = blob.download_as_string()
-    return json.loads(content)
-
-
-def download_image_from_gcs(gcs_uri: str) -> bytes:
-    """Download image bytes from GCS."""
-    if not gcs_uri.startswith("gs://"):
-        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
-
-    parts = gcs_uri[5:].split("/", 1)
-    bucket_name = parts[0]
-    blob_name = parts[1] if len(parts) > 1 else ""
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(blob_name)
-
-    return blob.download_as_bytes()
-
-
-def calculate_plddt_stats(plddt_scores: list) -> dict:
-    """Calculate pLDDT statistics from per-residue array."""
-    scores = np.array(plddt_scores)
-
-    stats = {
-        "mean": float(np.mean(scores)),
-        "median": float(np.median(scores)),
-        "min": float(np.min(scores)),
-        "max": float(np.max(scores)),
-        "std": float(np.std(scores)),
-        "per_residue": plddt_scores,
-    }
-
-    # Distribution by confidence band
-    distribution = {}
-    for min_val, max_val, label in PLDDT_BANDS:
-        count = int(np.sum((scores >= min_val) & (scores <= max_val)))
-        distribution[label] = count
-
-    stats["distribution"] = distribution
-    return stats
-
-
-def get_quality_assessment(plddt_mean: float) -> str:
-    """Get quality assessment based on mean pLDDT."""
-    if plddt_mean >= 90:
-        return "very_high_confidence"
-    elif plddt_mean >= 70:
-        return "high_confidence"
-    elif plddt_mean >= 50:
-        return "low_confidence"
-    else:
-        return "very_low_confidence"
 
 
 def parse_cif_chains(cif_text: str) -> list[dict]:
-    """Parse CIF atom_site records to extract per-chain info.
-
-    Returns list of dicts sorted by chain order:
-        [{'chain_id': 'A', 'atom_count': 601, 'residue_count': 76,
-          'comp_ids': {'MET','GLN',...}, 'molecule_type': 'protein'}, ...]
-    """
-    chains = {}  # chain_id -> {atoms, residues set, comp_ids set}
+    """Parse CIF atom_site records to extract per-chain info."""
+    chains = {}
     chain_order = []
 
     for line in cif_text.split("\n"):
@@ -186,7 +68,6 @@ def parse_cif_chains(cif_text: str) -> list[dict]:
     for cid in chain_order:
         info = chains[cid]
         comp_ids = info["comp_ids"]
-        # Classify molecule type from residue names
         standard_aa = {
             "ALA",
             "ARG",
@@ -234,37 +115,17 @@ def parse_cif_chains(cif_text: str) -> list[dict]:
     return result
 
 
-def download_text_from_gcs(gcs_uri: str) -> str:
-    """Download text file from GCS."""
-    if not gcs_uri.startswith("gs://"):
-        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
-    parts = gcs_uri[5:].split("/", 1)
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(parts[0])
-    blob = bucket.blob(parts[1] if len(parts) > 1 else "")
-    return blob.download_as_text()
-
-
 def plot_plddt(
     plddt_scores: list,
     sample_name: str,
     output_path: str,
     chain_info: list[dict] = None,
 ) -> None:
-    """Generate pLDDT per-atom plot with chain boundary annotations.
-
-    Args:
-        plddt_scores: Atom-level pLDDT scores
-        sample_name: Name for plot title
-        output_path: Local path to save plot PNG
-        chain_info: Chain info from parse_cif_chains() for boundary lines
-    """
+    """Generate pLDDT per-atom plot with chain boundary annotations."""
     scores = np.array(plddt_scores)
     fig, ax = plt.subplots(figsize=(12, 4))
-
     residues = np.arange(1, len(scores) + 1)
 
-    # If chain info available, color each chain differently
     if chain_info and len(chain_info) > 1:
         chain_colors = [
             "#1f77b4",
@@ -294,18 +155,15 @@ def plot_plddt(
                 label=label,
             )
 
-            # Draw chain boundary line
             if i < len(chain_info) - 1:
                 boundary = offset + n
                 ax.axvline(
                     x=boundary, color="gray", linestyle="--", alpha=0.5, linewidth=1
                 )
-
             offset += n
     else:
         ax.plot(residues, scores, linewidth=1.5, color="#1f77b4")
 
-    # Confidence band backgrounds
     ax.axhspan(90, 100, alpha=0.08, color="green")
     ax.axhspan(70, 90, alpha=0.08, color="yellow")
     ax.axhspan(50, 70, alpha=0.08, color="orange")
@@ -323,37 +181,30 @@ def plot_plddt(
     plt.tight_layout()
     plt.savefig(output_path, dpi=150, bbox_inches="tight")
     plt.close(fig)
-    logger.info(f"Generated pLDDT plot: {output_path}")
 
 
 def plot_pde(
     pde_matrix: list, sample_name: str, output_path: str, chain_info: list[dict] = None
 ) -> None:
-    """Generate PDE heatmap with chain boundary lines.
-
-    The PDE matrix is token-level (residue/entity level), not atom-level.
-    Chain boundaries are drawn using residue counts.
-    """
+    """Generate PDE heatmap with chain boundary lines."""
     pde = np.array(pde_matrix)
     fig, ax = plt.subplots(figsize=(8, 7))
 
     cmap = plt.cm.Greens_r
     im = ax.imshow(pde, vmin=0, vmax=np.max(pde), cmap=cmap)
 
-    # Draw chain boundary lines on heatmap
     if chain_info and len(chain_info) > 1:
         offset = 0
         for i, ci in enumerate(chain_info):
             n = ci["residue_count"]
             if i < len(chain_info) - 1:
-                boundary = offset + n - 0.5  # center on boundary
+                boundary = offset + n - 0.5
                 ax.axhline(
                     y=boundary, color="white", linestyle="-", linewidth=1.5, alpha=0.8
                 )
                 ax.axvline(
                     x=boundary, color="white", linestyle="-", linewidth=1.5, alpha=0.8
                 )
-                # Label chain
                 mid = offset + n / 2
                 ax.text(
                     -2,
@@ -381,7 +232,7 @@ def plot_pde(
 
     cbar = plt.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
     cbar.set_label(
-        "Predicted Distance Error (A)", rotation=270, labelpad=20, fontsize=10
+        "Predicted Distance Error (Å)", rotation=270, labelpad=20, fontsize=10
     )
 
     ax.set_xlabel("Scored Residue", fontsize=10)
@@ -395,19 +246,10 @@ def plot_pde(
     plt.tight_layout()
     plt.savefig(output_path, dpi=150, bbox_inches="tight")
     plt.close(fig)
-    logger.info(f"Generated PDE plot: {output_path}")
 
 
 def calculate_per_chain_plddt(plddt_scores: list, chain_info: list[dict]) -> dict:
-    """Calculate per-chain pLDDT statistics.
-
-    Args:
-        plddt_scores: Atom-level pLDDT scores
-        chain_info: Chain info from parse_cif_chains()
-
-    Returns:
-        Dict mapping chain_id to stats dict with mean, min, max, molecule_type, comp_ids
-    """
+    """Calculate per-chain pLDDT statistics."""
     scores = np.array(plddt_scores)
     result = {}
     offset = 0
@@ -438,16 +280,7 @@ def plot_iptm_matrix(
     chain_ptm: dict = None,
     chain_info: list = None,
 ) -> None:
-    """Generate chain x chain ipTM heatmap with per-chain pTM on diagonal.
-
-    Args:
-        chain_pair_iptm: Dict like {"(A, B)": 0.72}
-        sample_name: Name for plot title
-        output_path: Local path to save plot PNG
-        chain_ptm: Dict like {"A": 0.87, "B": 0.46} for diagonal values
-        chain_info: Chain info for molecule type labels
-    """
-    # Parse chain pairs and build matrix
+    """Generate chain x chain ipTM heatmap with per-chain pTM on diagonal."""
     chains = set()
     pairs = {}
     for key, value in chain_pair_iptm.items():
@@ -459,26 +292,22 @@ def plot_iptm_matrix(
             chains.add(c2)
             pairs[(c1, c2)] = value
 
-    # Also add chains from chain_ptm (for diagonal)
     if chain_ptm:
         chains.update(chain_ptm.keys())
 
     chains = sorted(chains)
     n = len(chains)
     if n < 2:
-        logger.info("Skipping ipTM matrix — fewer than 2 chains")
         return
 
     matrix = np.zeros((n, n))
     for i, c1 in enumerate(chains):
         for j, c2 in enumerate(chains):
             if i == j and chain_ptm and c1 in chain_ptm:
-                # Diagonal: per-chain pTM
                 matrix[i, j] = chain_ptm[c1]
             else:
                 matrix[i, j] = pairs.get((c1, c2), pairs.get((c2, c1), 0.0))
 
-    # Build labels with molecule type
     chain_labels = []
     ci_map = {c["chain_id"]: c for c in (chain_info or [])}
     for c in chains:
@@ -524,281 +353,6 @@ def plot_iptm_matrix(
     plt.tight_layout(pad=1.5)
     plt.savefig(output_path, dpi=150, bbox_inches="tight")
     plt.close(fig)
-    logger.info(f"Generated ipTM matrix plot: {output_path}")
-
-
-def main():
-    """Main entry point for Cloud Run Job task."""
-
-    task_index = int(os.getenv("CLOUD_RUN_TASK_INDEX", "0"))
-    task_count = int(os.getenv("CLOUD_RUN_TASK_COUNT", "1"))
-
-    logger.info(f"Task {task_index}/{task_count}: Starting")
-
-    bucket_name = os.getenv("GCS_BUCKET")
-    analysis_path = os.getenv("ANALYSIS_PATH")
-
-    if not bucket_name:
-        logger.error("GCS_BUCKET environment variable not set")
-        sys.exit(1)
-
-    if not analysis_path:
-        logger.error("ANALYSIS_PATH environment variable not set")
-        sys.exit(1)
-
-    logger.info(f"Bucket: {bucket_name}, Analysis path: {analysis_path}")
-
-    # Load task configuration
-    task_config_uri = f"{analysis_path}task_config.json"
-    logger.info(f"Task {task_index}: Loading configuration from {task_config_uri}")
-
-    try:
-        task_config = download_json_from_gcs(task_config_uri)
-    except Exception as e:
-        logger.error(f"Failed to load task configuration: {e}")
-        sys.exit(1)
-
-    predictions = task_config.get("predictions", [])
-    if task_index >= len(predictions):
-        logger.error(
-            f"Task index {task_index} out of range (total predictions: {len(predictions)})"
-        )
-        sys.exit(1)
-
-    pred = predictions[task_index]
-    job_id = task_config.get("job_id")
-
-    sample_name = pred["sample_name"]
-    cif_uri = pred["cif_uri"]
-    confidences_uri = pred["confidences_uri"]
-    aggregated_uri = pred["aggregated_uri"]
-    output_uri = pred["output_uri"]
-
-    logger.info(f"Starting analysis for sample: {sample_name}")
-    logger.info(f"Job ID: {job_id}")
-
-    try:
-        # Load aggregated confidences (summary scores)
-        logger.info("Loading aggregated confidences...")
-        aggregated = download_json_from_gcs(aggregated_uri)
-
-        # Load per-residue confidences
-        logger.info("Loading per-residue confidences...")
-        confidences = download_json_from_gcs(confidences_uri)
-
-        # Parse CIF to get chain info (atom counts, molecule types)
-        logger.info("Parsing CIF for chain info...")
-        chain_info = []
-        try:
-            cif_text = download_text_from_gcs(cif_uri)
-            chain_info = parse_cif_chains(cif_text)
-            for ci in chain_info:
-                logger.info(
-                    f"  Chain {ci['chain_id']}: {ci['molecule_type']}, "
-                    f"{ci['atom_count']} atoms, {ci['residue_count']} residues, "
-                    f"comps: {ci['comp_ids'][:5]}"
-                )
-        except Exception as e:
-            logger.warning(f"Could not parse CIF for chain info: {e}")
-
-        # Extract metrics from aggregated.
-        # OF3 uses AF3-style ranking: sample_ranking_score = 0.8*iptm + 0.2*ptm.
-        # For monomers iptm≈0 (no interface), so sample_ranking_score is near 0
-        # and meaningless for comparing prediction quality. Use ptm for monomers.
-        ptm = aggregated.get("ptm", 0.0)
-        iptm = aggregated.get("iptm", 0.0)
-        sample_ranking_score = aggregated.get(
-            "sample_ranking_score", aggregated.get("ranking_score", 0.0)
-        )
-        # Determine monomer vs complex from chain_info parsed from CIF
-        _is_monomer = len(chain_info) <= 1
-        if _is_monomer:
-            # pTM is the right quality metric for single-chain predictions
-            ranking_score = ptm if ptm > 0.0 else sample_ranking_score
-            logger.info(
-                f"Monomer: using ptm={ptm:.4f} as ranking_score "
-                f"(sample_ranking_score={sample_ranking_score:.4f} suppressed — iptm≈0 for monomers)"
-            )
-        else:
-            ranking_score = sample_ranking_score
-        avg_plddt = aggregated.get("avg_plddt", 0.0)
-        gpde = aggregated.get("gpde", None)
-        has_clash = aggregated.get("has_clash", 0)
-        disorder = aggregated.get("disorder", 0)
-        chain_ptm = aggregated.get("chain_ptm", {})
-        chain_pair_iptm = aggregated.get("chain_pair_iptm", {})
-
-        # Get per-atom pLDDT from confidences
-        plddt_scores = confidences.get("plddt", [])
-        pde_matrix = confidences.get("pde", None)
-
-        # Calculate overall pLDDT stats
-        if plddt_scores:
-            plddt_stats = calculate_plddt_stats(plddt_scores)
-        else:
-            plddt_stats = {
-                "mean": avg_plddt,
-                "median": avg_plddt,
-                "min": 0.0,
-                "max": 100.0,
-                "std": 0.0,
-                "per_residue": [],
-                "distribution": {},
-            }
-
-        # Calculate per-chain pLDDT stats (protein vs ligand breakdown)
-        per_chain_plddt = {}
-        if plddt_scores and chain_info:
-            per_chain_plddt = calculate_per_chain_plddt(plddt_scores, chain_info)
-            for cid, stats in per_chain_plddt.items():
-                logger.info(
-                    f"  Chain {cid} ({stats['molecule_type']}): "
-                    f"mean pLDDT={stats['mean']:.1f}"
-                )
-
-        quality = get_quality_assessment(plddt_stats["mean"])
-
-        # Generate plots
-        logger.info("Generating visualizations...")
-        plot_files = {}
-        output_base = output_uri.rsplit("/", 1)[0]
-
-        # pLDDT plot (with chain boundaries and per-chain coloring)
-        if plddt_scores:
-            plddt_plot_path = f"/tmp/plddt_plot_{task_index}.png"
-            plot_plddt(
-                plddt_scores,
-                sample_name,
-                plddt_plot_path,
-                chain_info=chain_info if chain_info else None,
-            )
-            plddt_plot_uri = f"{output_base}/plddt_plot_{task_index}.png"
-            upload_to_gcs(plddt_plot_path, plddt_plot_uri)
-            plot_files["plddt_plot"] = plddt_plot_uri
-            os.remove(plddt_plot_path)
-
-        # PDE heatmap with chain boundary lines
-        if pde_matrix:
-            pde_plot_path = f"/tmp/pde_plot_{task_index}.png"
-            plot_pde(
-                pde_matrix,
-                sample_name,
-                pde_plot_path,
-                chain_info=chain_info if chain_info else None,
-            )
-            pde_plot_uri = f"{output_base}/pde_plot_{task_index}.png"
-            upload_to_gcs(pde_plot_path, pde_plot_uri)
-            plot_files["pde_plot"] = pde_plot_uri
-            os.remove(pde_plot_path)
-
-        # ipTM matrix heatmap (multi-chain: needs at least 2 chains)
-        if chain_pair_iptm and len(chain_info) > 1:
-            iptm_plot_path = f"/tmp/iptm_matrix_{task_index}.png"
-            plot_iptm_matrix(
-                chain_pair_iptm,
-                sample_name,
-                iptm_plot_path,
-                chain_ptm=chain_ptm,
-                chain_info=chain_info,
-            )
-            if os.path.exists(iptm_plot_path):
-                iptm_plot_uri = f"{output_base}/iptm_matrix_{task_index}.png"
-                upload_to_gcs(iptm_plot_path, iptm_plot_uri)
-                plot_files["iptm_matrix_plot"] = iptm_plot_uri
-                os.remove(iptm_plot_path)
-
-        # Build analysis result
-        analysis = {
-            "prediction_index": task_index,
-            "sample_name": sample_name,
-            "rank": task_index + 1,
-            "ranking_score": ranking_score,
-            "ptm": ptm,
-            "iptm": iptm,
-            "avg_plddt": avg_plddt,
-            "gpde": gpde,
-            "has_clash": has_clash,
-            "disorder": disorder,
-            "chain_ptm": chain_ptm,
-            "chain_pair_iptm": chain_pair_iptm,
-            "per_chain_plddt": per_chain_plddt,
-            "chain_info": chain_info,
-            "plddt_mean": plddt_stats["mean"],
-            "plddt_median": plddt_stats["median"],
-            "plddt_min": plddt_stats["min"],
-            "plddt_max": plddt_stats["max"],
-            "plddt_std": plddt_stats["std"],
-            "plddt_distribution": plddt_stats["distribution"],
-            "plddt_scores": plddt_stats["per_residue"],
-            "quality_assessment": quality,
-            "has_pde": pde_matrix is not None,
-            "cif_uri": cif_uri,
-            "aggregated_uri": aggregated_uri,
-            "confidences_uri": confidences_uri,
-            "analyzed_at": datetime.now(timezone.utc).replace(tzinfo=None).isoformat()
-            + "Z",
-            "plots": plot_files,
-        }
-
-        # Write to JSON and upload
-        local_json = f"/tmp/analysis_{task_index}.json"
-        with open(local_json, "w") as f:
-            json.dump(analysis, f, indent=2)
-
-        upload_to_gcs(local_json, output_uri)
-        os.remove(local_json)
-
-        logger.info(f"Completed analysis for sample {sample_name}")
-        logger.info(f"   ranking_score: {ranking_score:.4f}")
-        logger.info(f"   pLDDT mean: {plddt_stats['mean']:.2f}")
-        logger.info(f"   Quality: {quality}")
-
-        # Last task triggers consolidation
-        if task_index == task_count - 1:
-            logger.info(
-                f"Last task ({task_index + 1}/{task_count}), waiting for all tasks..."
-            )
-
-            import time
-
-            max_wait = 120
-            wait_interval = 2
-            waited = 0
-
-            storage_client = storage.Client()
-            bucket_obj = storage_client.bucket(bucket_name)
-            analysis_prefix = analysis_path.replace(f"gs://{bucket_name}/", "")
-
-            while waited < max_wait:
-                blobs = bucket_obj.list_blobs(prefix=analysis_prefix)
-                completed_files = [
-                    b.name
-                    for b in blobs
-                    if "prediction_" in b.name and b.name.endswith("_analysis.json")
-                ]
-
-                if len(completed_files) >= task_count:
-                    logger.info(
-                        f"All {task_count} analysis files found, starting consolidation"
-                    )
-                    break
-
-                logger.info(
-                    f"Waiting... ({len(completed_files)}/{task_count} files found)"
-                )
-                time.sleep(wait_interval)
-                waited += wait_interval
-
-            try:
-                consolidate_results(job_id, analysis_path)
-            except Exception as e:
-                logger.error(f"Consolidation failed: {e}", exc_info=True)
-
-        sys.exit(0)
-
-    except Exception as e:
-        logger.error(f"Error analyzing sample {sample_name}: {e}", exc_info=True)
-        sys.exit(1)
 
 
 def generate_gemini_expert_analysis(summary_data: dict) -> dict:
@@ -808,7 +362,6 @@ def generate_gemini_expert_analysis(summary_data: dict) -> dict:
         summary = summary_data["summary"]
         top_preds = summary_data["top_predictions"][:3]
 
-        # Build chain composition description
         chain_desc = ""
         for ci in best.get("chain_info", []):
             if ci["molecule_type"] == "ligand":
@@ -857,7 +410,6 @@ def generate_gemini_expert_analysis(summary_data: dict) -> dict:
 - Disorder: {"Yes" if best.get("disorder") else "No"}
 """
 
-        # Add chain-level metrics
         if best.get("chain_ptm"):
             prompt += "\n**Per-Chain pTM:**\n"
             for chain, score in best["chain_ptm"].items():
@@ -868,7 +420,6 @@ def generate_gemini_expert_analysis(summary_data: dict) -> dict:
             for pair, score in best["chain_pair_iptm"].items():
                 prompt += f"- {pair}: {score:.4f}\n"
 
-        # Add per-chain pLDDT breakdown (protein vs ligand)
         if best.get("per_chain_plddt"):
             prompt += "\n**Per-Chain pLDDT (protein vs ligand breakdown):**\n"
             for cid, stats in best["per_chain_plddt"].items():
@@ -892,7 +443,6 @@ def generate_gemini_expert_analysis(summary_data: dict) -> dict:
             prompt += f"- Low (50-70): {dist.get('low_confidence', 0)} atoms\n"
             prompt += f"- Very Low (<50): {dist.get('very_low_confidence', 0)} atoms\n"
 
-        # Add ranking score statistics across all samples
         all_preds = summary_data.get("all_predictions_summary", [])
         if len(all_preds) > 1:
             all_scores = [p["ranking_score"] for p in all_preds]
@@ -905,7 +455,6 @@ def generate_gemini_expert_analysis(summary_data: dict) -> dict:
             clash_str = " [CLASH]" if pred.get("has_clash") else ""
             prompt += f"{i}. {pred['sample_name']}: ranking_score {pred['ranking_score']:.4f}, pTM {pred['ptm']:.4f}, ipTM {pred['iptm']:.4f}{clash_str}\n"
 
-        # Add input info if available
         if summary["protein_info"].get("fasta_sequence"):
             fasta_seq = summary["protein_info"]["fasta_sequence"]
             fasta_header = summary["protein_info"].get("fasta_header", "Unknown")
@@ -932,7 +481,6 @@ IMPORTANT: Begin your response directly with the analysis. Do NOT include any pr
 
         content_parts = [prompt]
 
-        # Include plot images if available
         if best.get("plots", {}).get("plddt_plot"):
             try:
                 img_uri = best["plots"]["plddt_plot"]
@@ -970,14 +518,10 @@ IMPORTANT: Begin your response directly with the analysis. Do NOT include any pr
             if "preview" in gemini_model
             else os.getenv("REGION", "us-central1")
         )
-        logger.info(
-            f"Initializing Google GenAI (Agent Platform): project={project_id}, location={location}, model={gemini_model}"
-        )
 
         with genai.Client(
             vertexai=True, project=project_id, location=location
         ) as client:
-            # Generate content using Gemini API
             response = client.models.generate_content(
                 model=gemini_model,
                 contents=content_parts,
@@ -1007,18 +551,14 @@ IMPORTANT: Begin your response directly with the analysis. Do NOT include any pr
         return {"status": "error", "error": str(e)}
 
 
-def consolidate_results(job_id: str, analysis_path: str):
-    """Consolidate per-sample analysis results into summary.json."""
-    import time
-
-    logger.info(f"Consolidating results from {analysis_path}")
+def consolidate_results(
+    job_id: str, analysis_path: str, bucket_name: str, task_count: int
+):
+    """Consolidate per-sample OF3 analysis results into summary.json."""
+    logger.info(f"Consolidating OF3 results from {analysis_path}")
     time.sleep(2)
 
-    if not analysis_path.startswith("gs://"):
-        raise ValueError(f"Invalid analysis_path format: {analysis_path}")
-
     parts = analysis_path[5:].split("/", 1)
-    bucket_name = parts[0]
     prefix = parts[1] if len(parts) > 1 else ""
 
     if prefix and not prefix.endswith("/"):
@@ -1026,7 +566,6 @@ def consolidate_results(job_id: str, analysis_path: str):
     if prefix.startswith("/"):
         prefix = prefix[1:]
 
-    # List all per-sample analysis files
     storage_client = storage.Client()
     bucket = storage_client.bucket(bucket_name)
 
@@ -1039,7 +578,6 @@ def consolidate_results(job_id: str, analysis_path: str):
 
     logger.info(f"Found {len(completed_files)} analysis files to consolidate")
 
-    # Read all analyses
     all_analyses = []
     for file_path in completed_files:
         try:
@@ -1052,7 +590,6 @@ def consolidate_results(job_id: str, analysis_path: str):
         logger.error("No analyses found to consolidate")
         return
 
-    # Get job metadata from Agent Platform
     all_labels = {}
     job_metadata = {"labels": {}, "parameters": {}}
     try:
@@ -1074,7 +611,6 @@ def consolidate_results(job_id: str, analysis_path: str):
         job = client.get_pipeline_job(request=request)
 
         all_labels = dict(job.labels) if hasattr(job, "labels") and job.labels else {}
-
         pipeline_parameters = {}
         if hasattr(job, "runtime_config") and hasattr(
             job.runtime_config, "parameter_values"
@@ -1109,24 +645,18 @@ def consolidate_results(job_id: str, analysis_path: str):
             job_metadata["duration_formatted"] = (
                 f"{hours}h {minutes}m" if hours > 0 else f"{minutes:.1f}m"
             )
-
     except Exception as e:
         logger.error(f"Could not get job metadata: {e}", exc_info=True)
 
-    # Sort by ranking_score (descending — higher is better)
     all_analyses.sort(key=lambda x: x.get("ranking_score", 0), reverse=True)
 
-    # Re-rank
     for i, analysis in enumerate(all_analyses):
         analysis["rank"] = i + 1
 
     best = all_analyses[0]
-
-    # Calculate summary statistics
     ranking_scores = [a["ranking_score"] for a in all_analyses]
     plddt_means = [a["plddt_mean"] for a in all_analyses]
 
-    # Try to get input sequence and raw query JSON
     fasta_sequence = None
     fasta_header = None
     input_query_json = None
@@ -1149,7 +679,6 @@ def consolidate_results(job_id: str, analysis_path: str):
                     ]
                     fasta_sequence = "/".join(seqs) if seqs else None
 
-                    # Build clean query JSON for resubmission (strip internal paths)
                     clean_queries = {}
                     for qname, qdata in queries.items():
                         clean_chains = []
@@ -1170,7 +699,6 @@ def consolidate_results(job_id: str, analysis_path: str):
     except Exception as e:
         logger.warning(f"Could not load input sequence: {e}")
 
-    # Build chain composition from best prediction's chain_info
     chain_composition = []
     best_chain_info = best.get("chain_info", [])
     total_residues = 0
@@ -1252,19 +780,9 @@ def consolidate_results(job_id: str, analysis_path: str):
         ],
     }
 
-    # Generate Gemini expert analysis
-    logger.info("Generating Gemini expert analysis...")
     expert_analysis = generate_gemini_expert_analysis(summary_data)
     summary_data["expert_analysis"] = expert_analysis
 
-    if expert_analysis["status"] == "success":
-        logger.info("Expert analysis generated successfully")
-    else:
-        logger.error(
-            f"Expert analysis failed: {expert_analysis.get('error', 'Unknown')}"
-        )
-
-    # Write summary to GCS
     summary_uri = f"{analysis_path}summary.json"
     local_summary = "/tmp/summary.json"
 
@@ -1275,11 +793,184 @@ def consolidate_results(job_id: str, analysis_path: str):
     os.remove(local_summary)
 
     logger.info(f"Consolidation complete! Summary written to {summary_uri}")
-    logger.info(
-        f"   Best sample: {best['sample_name']} (ranking_score: {best['ranking_score']:.4f})"
-    )
-    logger.info(f"   Labels captured: {list(all_labels.keys())}")
 
 
-if __name__ == "__main__":
-    main()
+def run_task(
+    task_index: int,
+    task_count: int,
+    task_config: dict,
+    bucket_name: str,
+    analysis_path: str,
+):
+    """Execute OpenFold3 predictions analysis task."""
+    predictions = task_config.get("predictions", [])
+    if task_index >= len(predictions):
+        logger.error(f"Task index {task_index} out of range")
+        sys.exit(1)
+
+    pred = predictions[task_index]
+    job_id = task_config.get("job_id")
+
+    sample_name = pred["sample_name"]
+    cif_uri = pred["cif_uri"]
+    confidences_uri = pred["confidences_uri"]
+    aggregated_uri = pred["aggregated_uri"]
+    output_uri = pred["output_uri"]
+
+    logger.info(f"Starting OF3 analysis for sample: {sample_name}")
+
+    try:
+        aggregated = download_json_from_gcs(aggregated_uri)
+        confidences = download_json_from_gcs(confidences_uri)
+
+        chain_info = []
+        try:
+            cif_text = download_text_from_gcs(cif_uri)
+            chain_info = parse_cif_chains(cif_text)
+        except Exception as e:
+            logger.warning(f"Could not parse CIF: {e}")
+
+        ptm = aggregated.get("ptm", 0.0)
+        iptm = aggregated.get("iptm", 0.0)
+        sample_ranking_score = aggregated.get(
+            "sample_ranking_score", aggregated.get("ranking_score", 0.0)
+        )
+
+        _is_monomer = len(chain_info) <= 1
+        if _is_monomer:
+            ranking_score = ptm if ptm > 0.0 else sample_ranking_score
+        else:
+            ranking_score = sample_ranking_score
+
+        avg_plddt = aggregated.get("avg_plddt", 0.0)
+        gpde = aggregated.get("gpde", None)
+        has_clash = aggregated.get("has_clash", 0)
+        disorder = aggregated.get("disorder", 0)
+        chain_ptm = aggregated.get("chain_ptm", {})
+        chain_pair_iptm = aggregated.get("chain_pair_iptm", {})
+
+        plddt_scores = confidences.get("plddt", [])
+        pde_matrix = confidences.get("pde", None)
+
+        if plddt_scores:
+            plddt_stats = calculate_plddt_stats(plddt_scores)
+        else:
+            plddt_stats = {
+                "mean": avg_plddt,
+                "median": avg_plddt,
+                "min": 0.0,
+                "max": 100.0,
+                "std": 0.0,
+                "per_residue": [],
+                "distribution": {},
+            }
+
+        per_chain_plddt = {}
+        if plddt_scores and chain_info:
+            per_chain_plddt = calculate_per_chain_plddt(plddt_scores, chain_info)
+
+        quality = get_quality_assessment(plddt_stats["mean"])
+
+        # Generate plots
+        plot_files = {}
+        output_base = output_uri.rsplit("/", 1)[0]
+
+        if plddt_scores:
+            plddt_plot_path = f"/tmp/plddt_plot_{task_index}.png"
+            plot_plddt(plddt_scores, sample_name, plddt_plot_path, chain_info)
+            plddt_plot_uri = f"{output_base}/plddt_plot_{task_index}.png"
+            upload_to_gcs(plddt_plot_path, plddt_plot_uri)
+            plot_files["plddt_plot"] = plddt_plot_uri
+            os.remove(plddt_plot_path)
+
+        if pde_matrix:
+            pde_plot_path = f"/tmp/pde_plot_{task_index}.png"
+            plot_pde(pde_matrix, sample_name, pde_plot_path, chain_info)
+            pde_plot_uri = f"{output_base}/pde_plot_{task_index}.png"
+            upload_to_gcs(pde_plot_path, pde_plot_uri)
+            plot_files["pde_plot"] = pde_plot_uri
+            os.remove(pde_plot_path)
+
+        if chain_pair_iptm and len(chain_info) > 1:
+            iptm_plot_path = f"/tmp/iptm_matrix_{task_index}.png"
+            plot_iptm_matrix(
+                chain_pair_iptm,
+                sample_name,
+                iptm_plot_path,
+                chain_ptm,
+                chain_info,
+            )
+            if os.path.exists(iptm_plot_path):
+                iptm_plot_uri = f"{output_base}/iptm_matrix_{task_index}.png"
+                upload_to_gcs(iptm_plot_path, iptm_plot_uri)
+                plot_files["iptm_matrix_plot"] = iptm_plot_uri
+                os.remove(iptm_plot_path)
+
+        analysis = {
+            "prediction_index": task_index,
+            "sample_name": sample_name,
+            "rank": task_index + 1,
+            "ranking_score": ranking_score,
+            "ptm": ptm,
+            "iptm": iptm,
+            "avg_plddt": avg_plddt,
+            "gpde": gpde,
+            "has_clash": has_clash,
+            "disorder": disorder,
+            "chain_ptm": chain_ptm,
+            "chain_pair_iptm": chain_pair_iptm,
+            "per_chain_plddt": per_chain_plddt,
+            "chain_info": chain_info,
+            "plddt_mean": plddt_stats["mean"],
+            "plddt_median": plddt_stats["median"],
+            "plddt_min": plddt_stats["min"],
+            "plddt_max": plddt_stats["max"],
+            "plddt_std": plddt_stats["std"],
+            "plddt_distribution": plddt_stats["distribution"],
+            "plddt_scores": plddt_stats["per_residue"],
+            "quality_assessment": quality,
+            "has_pde": pde_matrix is not None,
+            "cif_uri": cif_uri,
+            "aggregated_uri": aggregated_uri,
+            "confidences_uri": confidences_uri,
+            "analyzed_at": datetime.now(timezone.utc).replace(tzinfo=None).isoformat()
+            + "Z",
+            "plots": plot_files,
+        }
+
+        local_json = f"/tmp/analysis_{task_index}.json"
+        with open(local_json, "w") as f:
+            json.dump(analysis, f, indent=2)
+
+        upload_to_gcs(local_json, output_uri)
+        os.remove(local_json)
+
+        # Consolidation check
+        if task_index == task_count - 1:
+            max_wait = 120
+            wait_interval = 2
+            waited = 0
+            storage_client = storage.Client()
+            bucket_obj = storage_client.bucket(bucket_name)
+            analysis_prefix = analysis_path.replace(f"gs://{bucket_name}/", "")
+
+            while waited < max_wait:
+                blobs = bucket_obj.list_blobs(prefix=analysis_prefix)
+                completed_files = [
+                    b.name
+                    for b in blobs
+                    if "prediction_" in b.name and b.name.endswith("_analysis.json")
+                ]
+
+                if len(completed_files) >= task_count:
+                    logger.info("All task analysis files found, starting consolidation")
+                    break
+
+                time.sleep(wait_interval)
+                waited += wait_interval
+
+            consolidate_results(job_id, analysis_path, bucket_name, task_count)
+
+    except Exception as e:
+        logger.error(f"Error running OF3 analysis task: {e}", exc_info=True)
+        raise
