+++ applications/foldrun/src/foldrun-analysis-job/requirements.txt
@@ -4,3 +4,6 @@ google-genai>=1.0.0
 numpy>=2.2.3
 matplotlib>=3.10.0
 seaborn>=0.13.2
+pyyaml>=6.0
+jax[cpu]==0.4.38
+dm-tree>=0.1.9
