+++ applications/foldrun/src/foldrun-analysis-job/shared_utils.py
@@ -1,0 +1,141 @@
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared GCS, mathematical, and assessment utilities for FoldRun prediction analysis."""
+
+import json
+import logging
+import numpy as np
+from google.cloud import storage
+
+logger = logging.getLogger(__name__)
+
+PLDDT_BANDS = [
+    (0, 50, "very_low_confidence"),
+    (50, 70, "low_confidence"),
+    (70, 90, "high_confidence"),
+    (90, 100, "very_high_confidence"),
+]
+
+
+def download_from_gcs(gcs_uri: str, local_path: str) -> None:
+    """Download file from GCS."""
+    if not gcs_uri.startswith("gs://"):
+        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
+
+    parts = gcs_uri[5:].split("/", 1)
+    bucket_name = parts[0]
+    blob_name = parts[1] if len(parts) > 1 else ""
+
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(blob_name)
+    blob.reload()
+    blob.download_to_filename(local_path)
+
+    size_mb = blob.size / 1024 / 1024 if blob.size else 0
+    logger.info(f"Downloaded {gcs_uri} ({size_mb:.2f} MB)")
+
+
+def upload_to_gcs(local_path: str, gcs_uri: str) -> None:
+    """Upload file to GCS."""
+    if not gcs_uri.startswith("gs://"):
+        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
+
+    parts = gcs_uri[5:].split("/", 1)
+    bucket_name = parts[0]
+    blob_name = parts[1] if len(parts) > 1 else ""
+
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(blob_name)
+    blob.upload_from_filename(local_path)
+    logger.info(f"Uploaded {local_path} to {gcs_uri}")
+
+
+def download_json_from_gcs(gcs_uri: str) -> dict:
+    """Download and parse JSON file from GCS."""
+    if not gcs_uri.startswith("gs://"):
+        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
+
+    parts = gcs_uri[5:].split("/", 1)
+    bucket_name = parts[0]
+    blob_name = parts[1] if len(parts) > 1 else ""
+
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(blob_name)
+    content = blob.download_as_string()
+    return json.loads(content)
+
+
+def download_text_from_gcs(gcs_uri: str) -> str:
+    """Download text file from GCS."""
+    if not gcs_uri.startswith("gs://"):
+        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
+    parts = gcs_uri[5:].split("/", 1)
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(parts[0])
+    blob = bucket.blob(parts[1] if len(parts) > 1 else "")
+    return blob.download_as_text()
+
+
+def download_image_from_gcs(gcs_uri: str) -> bytes:
+    """Download image bytes from GCS."""
+    if not gcs_uri.startswith("gs://"):
+        raise ValueError(f"Invalid GCS URI: {gcs_uri}")
+
+    parts = gcs_uri[5:].split("/", 1)
+    bucket_name = parts[0]
+    blob_name = parts[1] if len(parts) > 1 else ""
+
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(blob_name)
+    return blob.download_as_bytes()
+
+
+def calculate_plddt_stats(plddt_scores: list) -> dict:
+    """Calculate pLDDT statistics from per-residue array."""
+    scores = np.array(plddt_scores)
+
+    stats = {
+        "mean": float(np.mean(scores)),
+        "median": float(np.median(scores)),
+        "min": float(np.min(scores)),
+        "max": float(np.max(scores)),
+        "std": float(np.std(scores)),
+        "per_residue": plddt_scores,
+    }
+
+    # Distribution by confidence band
+    distribution = {}
+    for min_val, max_val, label in PLDDT_BANDS:
+        count = int(np.sum((scores >= min_val) & (scores <= max_val)))
+        distribution[label] = count
+
+    stats["distribution"] = distribution
+    return stats
+
+
+def get_quality_assessment(plddt_mean: float) -> str:
+    """Get quality assessment based on mean pLDDT."""
+    if plddt_mean >= 90:
+        return "very_high_confidence"
+    elif plddt_mean >= 70:
+        return "high_confidence"
+    elif plddt_mean >= 50:
+        return "low_confidence"
+    else:
+        return "very_low_confidence"
