--- applications/foldrun/src/of3-analysis-job/Dockerfile
@@ -1,53 +1,0 @@
-# Copyright 2026 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# OpenFold3 Analysis Job - Cloud Run Job Dockerfile
-
-# Stage 1: Build environment
-FROM python:3.13-slim AS builder
-
-# Set working directory
-WORKDIR /app
-
-# Install build dependencies
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    build-essential \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements and install dependencies
-COPY requirements.txt .
-RUN pip install --no-cache-dir --prefix=/install -r requirements.txt
-
-# Stage 2: Runtime image
-FROM python:3.13-slim
-
-# Set working directory
-WORKDIR /app
-
-# Copy installed dependencies from builder
-COPY --from=builder /install /usr/local
-
-# Copy application files
-COPY main.py .
-
-# Create non-root user for security
-RUN useradd -m -u 1000 appuser && chown -R appuser:appuser /app
-USER appuser
-
-# Set environment variables
-ENV PYTHONUNBUFFERED=1
-
-# Entry point
-ENTRYPOINT ["python", "main.py"]
-
