--- applications/foldrun/src/of3-analysis-job/cloudbuild.yaml
@@ -1,7 +1,0 @@
-steps:
-- name: 'gcr.io/cloud-builders/docker'
-  args: ['build', '-t', '${_IMAGE_PATH}', '.']
-images:
-- '${_IMAGE_PATH}'
-options:
-  logging: CLOUD_LOGGING_ONLY
